+++ src/CMakeLists.txt
@@ -1,4 +1,5 @@
 add_subdirectory(traceLog)
+add_subdirectory(types)
 add_subdirectory(ui)
 add_subdirectory(app)
 add_subdirectory(core)
