+++ src/core/CMakeLists.txt
@@ -27,7 +27,7 @@ target_include_directories(${NAME} PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
 target_compile_features(${NAME} PRIVATE cxx_std_20)
 
 target_link_libraries(${NAME}
-  PUBLIC Ui
+  PUBLIC Ui Types
   PRIVATE project_warnings
   PUBLIC TraceLog Image SDL2::SDL2 imgui Settings Camera)
