+++ src/ui/Ui/src/AppUi.cpp
@@ -1,5 +1,5 @@
 #include "AppUi.h"
-#include "Core/DataModel.h"
+#include "Types/DataModel.h"
 
 #include "imgui.h"
